test ImageCommentsMapper API (will be used in composition root with `RemoteLoader`)
